+++ yash-semantics/src/expansion/initial/arith.rs
@@ -23,6 +23,7 @@ use super::super::ErrorCause;
 use super::Env;
 use super::Error;
 use crate::expansion::expand_text;
+use std::convert::Infallible;
 use std::ops::Range;
 use std::rc::Rc;
 use yash_arith::eval;
@@ -133,7 +134,7 @@ impl ArithError {
 /// It is used to reproduce a location contained in the error cause.
 #[must_use]
 pub fn convert_error_cause(
-    cause: yash_arith::ErrorCause<ReadOnlyError>,
+    cause: yash_arith::ErrorCause<Infallible, ReadOnlyError>,
     source: &Rc<Code>,
 ) -> ErrorCause {
     use ArithError::*;
@@ -181,6 +182,7 @@ pub fn convert_error_cause(
             }
             yash_arith::EvalError::ReverseShifting => ErrorCause::ArithError(ReverseShifting),
             yash_arith::EvalError::AssignmentToValue => ErrorCause::ArithError(AssignmentToValue),
+            yash_arith::EvalError::GetVariableError(e) => match e {},
             yash_arith::EvalError::AssignVariableError(e) => ErrorCause::AssignReadOnly(e),
         },
     }
@@ -193,14 +195,15 @@ struct VarEnv<'a> {
 }
 
 impl<'a> yash_arith::Env for VarEnv<'a> {
+    type GetVariableError = Infallible;
     type AssignVariableError = ReadOnlyError;
 
     #[rustfmt::skip]
-    fn get_variable(&self, name: &str) -> Option<&str> {
+    fn get_variable(&self, name: &str) -> Result<Option<&str>, Infallible> {
         if let Some(Variable { value: Some(Scalar(value)), .. }) = self.env.variables.get(name) {
-            Some(value)
+            Ok(Some(value))
         } else {
-            None
+            Ok(None)
         }
     }